filter content out that is currently unhandled
I had a look at sync.cpp and checked which parts of the sync response
are currently handled and which not.
As I think it is unneccessary to let the unhandled data be transmitted
without beeing handled I added these filters.

In the same term I increased the timeout server-side to 30s as Riot
defaults to this value as well.
Especially now when a lots of presence-updates are not send anymore
this value is more relevant

Signed-Off-by: Matthias Kesler <krombel@krombel.de>