add sync filter for initial sync
To not add the default filter to several places a variable filter_
got introduced to store the filter.
It is now also possible to use a filter that is defined in
`client/sync_filter`. Advanced users might want to set an own filter
here.
Bit this is mainly done to prepare uploading the filter to the server
and just use a filter id that is stored on this location.
The uploading task is not done (yet)